requirements.txt
# Core data & ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.1.0

# API framework
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Serialization for model saving
joblib==1.4.2

# For JSON and HTTP handling
pydantic==2.7.1

# Optional but recommended for development
python-dotenv==1.0.1

mlflow==2.12.1
optuna==3.6.1

setuptools

evidently

pandera

openai

# Testing dependencies
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
